[script.tvguide] 2.0.1

+++ script.tvguide/source.py
@@ -108,6 +108,7 @@ class Database(object):
         self.updateInProgress = False
         self.updateFailed = False
         self.sourceNotConfigured = False
+        self.settingsChanged = None
 
         #buggalo.addExtraData('source', self.source.KEY)
         #for key in SETTINGS_TO_CHECK:
@@ -220,7 +221,8 @@ class Database(object):
     def _close(self):
         try:
             # rollback any non-commit'ed changes to avoid database lock
-            self.conn.rollback()
+            if self.conn:
+                self.conn.rollback()
         except sqlite3.OperationalError:
             pass # no transaction is active
         self.conn.close()
@@ -556,17 +558,18 @@ class Database(object):
         c.close()
         return expired
 
-
     def setCustomStreamUrl(self, channel, stream_url):
-        self._invokeAndBlockForResult(self._setCustomStreamUrl, channel, stream_url)
+        if stream_url is not None:
+            self._invokeAndBlockForResult(self._setCustomStreamUrl, channel, stream_url)
         # no result, but block until operation is done
 
     def _setCustomStreamUrl(self, channel, stream_url):
-        c = self.conn.cursor()
-        c.execute("DELETE FROM custom_stream_url WHERE channel=?", [channel.id])
-        c.execute("INSERT INTO custom_stream_url(channel, stream_url) VALUES(?, ?)", [channel.id, stream_url.decode('utf-8', 'ignore')])
-        self.conn.commit()
-        c.close()
+        if stream_url is not None:
+            c = self.conn.cursor()
+            c.execute("DELETE FROM custom_stream_url WHERE channel=?", [channel.id])
+            c.execute("INSERT INTO custom_stream_url(channel, stream_url) VALUES(?, ?)", [channel.id, stream_url.decode('utf-8', 'ignore')])
+            self.conn.commit()
+            c.close()
 
     def getCustomStreamUrl(self, channel):
         return self._invokeAndBlockForResult(self._getCustomStreamUrl, channel)
@@ -812,13 +815,8 @@ class XMLTVSource(Source):
         return parseXMLTV(context, f, f.size, self.logoFolder, progress_callback)
 
     def isUpdated(self, lastUpdated):
-        if hasattr(xbmcvfs, 'Stat'):
-            stat = xbmcvfs.Stat(self.xmltvFile)
-            mtime = stat.st_mtime()
-        else:
-            mode, ino, dev, nlink, uid, gid, size, atime, mtime, ctime = os.stat(self.xmltvFile)
-
-        fileUpdated = datetime.datetime.fromtimestamp(mtime)
+        stat = xbmcvfs.Stat(self.xmltvFile)
+        fileUpdated = datetime.datetime.fromtimestamp(stat.st_mtime())
         return fileUpdated > lastUpdated
 
 
@@ -887,19 +885,11 @@ def parseXMLTV(context, f, size, logoFolder, progress_callback):
         root.clear()
     f.close()
 
+
 class FileWrapper(object):
     def __init__(self, filename):
-        if hasattr(xbmcvfs, 'File'):
-            #xbmcvfs.File() was added in Frodo
-            self.vfsfile = xbmcvfs.File(filename)
-            self.size = self.vfsfile.size()
-        else:
-            print "xbmcvfs.File() is missing - perhaps you are running XBMC Eden? - retrying with python file opener"
-            try:
-                self.vfsfile = open(filename)
-            except IOError:
-                xbmcgui.Dialog().ok(strings(LOAD_ERROR_TITLE), 'smb://, nfs://, etc. is not support in Eden', 'Mount these on a system level instead. Filename:', filename)
-            self.size = os.path.getsize(filename)
+        self.vfsfile = xbmcvfs.File(filename)
+        self.size = self.vfsfile.size()
         self.bytesRead = 0
 
     def close(self):
@@ -911,6 +901,7 @@ class FileWrapper(object):
 
     def tell(self):
         return self.bytesRead
+
 
 def instantiateSource():
     SOURCES = {
